+++ packages/cli/CHANGELOG.md
@@ -5,6 +5,13 @@
 ### Patch Changes
 
 - 861aa5e: Add support of debug vite properties, add devtools for smart-clipping, security patches
+  - @hexajs-dev/common@0.9.1
+
+## 0.9.1
+
+### Patch Changes
+
+- 861aa5e: Add support of debug vite properties, add devtools for smart-clipping, security patches
   - @hexajs/common@0.9.1
 
 ## 2.0.0